fix: 提取重名函数
